external/robotouille/requirements.txt
altgraph==0.17.4
annotated-types==0.7.0
anthropic==0.34.2
antlr4-python3-runtime==4.9.3
anyio==4.4.0
appnope==0.1.4
asttokens==2.4.1
cachetools==5.5.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==2.2.1
comm==0.2.2
contourpy==1.0.7
cycler==0.11.0
debugpy==1.8.5
decorator==5.1.1
distro==1.9.0
exceptiongroup==1.2.2
executing==2.1.0
filelock==3.16.1
fonttools==4.38.0
fsspec==2024.9.0
google-ai-generativelanguage==0.6.9
google-api-core==2.19.2
google-api-python-client==2.145.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.1
googleapis-common-protos==1.65.0
grpcio==1.66.1
grpcio-status==1.66.1
gym==0.26.2
gym-notices==0.0.8
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.29.1
hydra-core==1.3.2
idna==3.7
imageio==2.25.1
importlib-metadata==6.0.0
importlib-resources==5.12.0
ipykernel==6.29.5
ipython==8.18.1
jedi==0.19.1
Jinja2==3.1.5
jiter==0.5.0
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.4
lorem-text==2.1
macholib==1.16.3
MarkupSafe==3.0.2
matplotlib==3.7.0
matplotlib-inline==0.1.7
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.0
numpy==1.24.2
omegaconf==2.3.0
openai==1.52.0
opencv-python==4.10.0.84
packaging==23.0
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
Pillow==9.4.0
platformdirs==4.3.3
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==5.28.1
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
pygame==2.1.3
pygame-ce==2.5.2
pygame_gui==0.6.12
Pygments==2.18.0
pyinstaller==6.11.0
pyinstaller-hooks-contrib==2024.9
pyparsing==3.0.9
python-dateutil==2.8.2
python-i18n==0.3.9
pytz==2024.2
PyWavelets==1.4.1
PyYAML==6.0.1
pyzmq==26.2.0
regex==2024.5.15
requests==2.32.3
rsa==4.9
safetensors==0.5.2
scikit-image==0.19.3
scipy==1.10.0
seaborn==0.13.2
six==1.16.0
sniffio==1.3.1
stack-data==0.6.3
sympy==1.13.1
tifffile==2023.2.3
tiktoken==0.7.0
tokenizers==0.21.0
torch==2.6.0
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
transformers==4.49.0
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
vllm==0.8.2
wcwidth==0.2.13
zipp==3.14.0
